+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+2.2.0
+-----
+
+* dropped support for PHP < 7.1
+
 2.1.0
 -----